fix timer and deadlock issues on linux
using signal for the timer prevents proper locking, as the
execution is replaced with the signal handler, and waiting for a
lock will cause a deadlock, since the original code isnt run
anymore.

using a proper sdl thread fixes the issue.
this code is not cleaned up, and should serve as a discussion base.