-Fix(#93) : Don't use signal() to do the work
Doesn't change anything to WIN32 code which works perfectly !

It is unsafe to call most function in a signal handler : So
we only write to a volatile sig_atomic_t variable, and then
use SleepIdle() to do the real work.
Warnings are issued when some "Interrupts" are lost.

pause() replaces usleep() but waits for the signal to be
triggered. That should ensure that Timer code is executed
as soon as the signal handler returns.
When this is not the case, Warnings are issued.

Code changes are minimal, I hope that makes them easily understandable.