timer: avoid Linux misfeature
The behaviour of setitimer(timer, NULL, &old_value) is unspecified. FreeBSD and other systems redirect it as getitimer(timer, &old_value), with a null pointer dereference in this specific case, as &old_value is also NULL. Other systems return an error. Linux seems to be the only one to treat that as intended here: as if called with a zero filled itimerval to disable the timer, but even Linux warns about and against that use.